Fix typecheck. extract queries to repositories
