+++ packages/cli/src/modules/insights/repositories/insights-raw.repository.ts
@@ -8,4 +8,18 @@ export class InsightsRawRepository extends Repository<InsightsRaw> {
 	constructor(dataSource: DataSource) {
 		super(InsightsRaw, dataSource.manager);
 	}
+
+	getRawInsightsBatchQuery(compactionBatchSize: number) {
+		// Build the query to gather raw insights data for the batch
+		const batchQuery = this.createQueryBuilder()
+			.select(
+				['id', 'metaId', 'type', 'value'].map((fieldName) =>
+					this.manager.connection.driver.escape(fieldName),
+				),
+			)
+			.addSelect('timestamp', 'periodStart')
+			.orderBy('timestamp', 'ASC')
+			.limit(compactionBatchSize);
+		return batchQuery;
+	}
 }
